simpleblog1/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.12.1
click==8.1.7
colorama==0.4.6
Django==4.2.9
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
